Create Sniffdogz Smart DCA Hedge
Thanks to xops37 for this open source pinescript

This script is a mainly a concept inspired by countless overpriced MT5 EAs and not meant to be used for trading as it randomly enters a trade, its to demonstrate how a hedging system could work on trading view, since trading view doesn't allow multiple trades in opposite directions simultaneously. The script randomly enters Short trade, if the TP is not hit it opens a Long  trade were there would normal by a SL. If keeps doing this with increasing Lot sizes until the TP is hit or the account is out of equity. Works best for directional markets as sideways markets can blow your account. This concept has alot of room for improvement ideally the recovery zone in the script should not be static and based on a indicator like Support and Resistance, same with the enteries.